Change vizro.figures.library to be original function definition
