+++ vizro-core/src/vizro/figures/__init__.py
@@ -1,3 +1,8 @@
-from vizro.figures._kpi_cards import kpi_card, kpi_card_reference
+"""Contains KPI card functions."""
 
-__all__ = ["kpi_card", "kpi_card_reference"]
+from vizro.models.types import capture
+
+from .library import kpi_card, kpi_card_reference
+
+kpi_card = capture("figure")(kpi_card)
+kpi_card_reference = capture("figure")(kpi_card_reference)
